DoorFaceLock-2.0/requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cmake==3.22.2
colorama==0.4.4
dlib @ file:///D:/bld/dlib_1617441823598/work
face-recognition==1.3.0
face-recognition-models==0.3.0
mysql-connector-python==8.0.27
numpy @ file:///D:/bld/numpy_1643009390959/work
opencv-python==4.5.4.60
Pillow==8.4.0
protobuf==3.19.1
pycparser==2.21
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
six==1.16.0
wincertstore==0.2
